pyproject.toml
[project]
name = "megabonk-builds"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.0",
    "alembic>=1.17.2",
    "bcrypt>=5.0.0",
    "fastapi>=0.124.4",
    "pydantic-settings>=2.12.0",
    "pyjwt[crypto]>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]
